Removed Old DAL
