+++ Graduate/Graduate.Droid/packages.config
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="sqlite-net-pcl" version="1.2.0" targetFramework="monoandroid70" />
-  <package id="SQLitePCLRaw.bundle_green" version="1.0.1" targetFramework="monoandroid70" />
-  <package id="SQLitePCLRaw.core" version="1.0.1" targetFramework="monoandroid70" />
-  <package id="SQLitePCLRaw.lib.e_sqlite3.android" version="1.0.1" targetFramework="monoandroid70" />
-  <package id="SQLitePCLRaw.provider.e_sqlite3.android" version="1.0.1" targetFramework="monoandroid70" />
+  <package id="SQLitePCLRaw.bundle_green" version="1.1.0" targetFramework="monoandroid70" />
+  <package id="SQLitePCLRaw.core" version="1.1.0" targetFramework="monoandroid70" />
+  <package id="SQLitePCLRaw.lib.e_sqlite3.android" version="1.1.0" targetFramework="monoandroid70" />
+  <package id="SQLitePCLRaw.provider.e_sqlite3.android" version="1.1.0" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.Animated.Vector.Drawable" version="24.2.1" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.Compat" version="24.2.1" targetFramework="monoandroid70" />
   <package id="Xamarin.Android.Support.Core.UI" version="24.2.1" targetFramework="monoandroid70" />
