--- Graduate/Graduate/DAL/SQLiteDB.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using SQLite;
-
-namespace Graduate.Core.DAL
-{
-    public class SQLiteDB
-    {
-        public SQLiteConnection DbConnection(String conn)
-        {
-            return new SQLiteConnection(conn);
-        }
-    }
-}
