+++ Graduate/Graduate/Data/DataAccessLayer/GraduateDatabase.cs
@@ -19,11 +19,14 @@ namespace Graduate.Core.Data.DataAccessLayer
             connection.DropTable<SchoolYear>();
             connection.DropTable<Semester>();
             connection.DropTable<Class>();
+            connection.DropTable<Grade>();
+            
 
             connection.CreateTable<SchoolYear>();
             connection.CreateTable<Semester>();
             connection.CreateTable<Class>();
-            
+            connection.CreateTable<Grade>();
+
         }
