+++ Graduate/Graduate/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="sqlite-net-pcl" version="1.2.0" targetFramework="portable45-net45+win8+wp8+wpa81" />
-  <package id="SQLitePCLRaw.bundle_green" version="1.0.1" targetFramework="portable45-net45+win8+wp8+wpa81" />
-  <package id="SQLitePCLRaw.core" version="1.0.1" targetFramework="portable45-net45+win8+wp8+wpa81" />
+  <package id="SQLitePCLRaw.bundle_green" version="1.1.0" targetFramework="portable45-net45+win8+wp8+wpa81" />
+  <package id="SQLitePCLRaw.core" version="1.1.0" targetFramework="portable45-net45+win8+wp8+wpa81" />
 </packages>
